devices/mirror/requirements.txt
# Mirror device runtime dependencies
# Install on the Mirror Raspberry Pi.

openai>=1.65.0
opencv-python>=4.10.0.84
Pillow>=11.1.0
PyYAML>=6.0.2
adafruit-circuitpython-servokit>=1.3.18
pytweening>=1.2.0
websockets>=12.0
